main-svc to main

+++ stat-service/dto/src/main/java/ru/practicum/ViewStats.java
@@ -10,5 +10,5 @@ import lombok.ToString;
 public class ViewStats {
     private String app;
     private String uri;
-    private int hits;
+    private Long hits;
 }
